refactoring merge code to be in mergeutils for reuse in other modules
