+++ setup.py
@@ -29,6 +29,7 @@ setup(name=NAME,
               'dumpSTR=dumpSTR.dumpSTR:main',
               'mergeSTR=mergeSTR.mergeSTR:main',
               'statSTR=statSTR.statSTR:main',
+              'compareSTR=compareSTR.compareSTR:main'
           ],
       },
       install_requires=['argparse',
