+++ test/test_mergeSTR/test_mergeSTR.py
@@ -1,5 +1,6 @@
 import argparse
 import os, sys
+import numpy as np
 import pytest
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), '..','..','mergeSTR'))
 from mergeSTR import * 
@@ -29,21 +30,6 @@ class DummyRecord:
         self.REF = ref
         self.ALTS = alts
         self.INFO = info
-
-
-# Test no such file or directory
-def test_WrongFile():
-    # Try a dummy file name. Make sure it doesn't exist before we try
-    fname1 = os.path.join(MRGVCFDIR, "test_non_existent1.vcf.gz")
-    fname2 = os.path.join(MRGVCFDIR, "test_non_existent2.vcf.gz")
-    if os.path.exists(fname1):
-        os.remove(fname1)
-    if os.path.exists(fname2):
-        os.remove(fname2)
-    print (os.path.isfile(fname1))
-    with pytest.raises(ValueError) as info:
-        LoadReaders([fname1, fname2])
-    assert "Could not find VCF file" in str(info.value)
 
 # Test right files or directory - GangSTR
 def test_GangSTRRightFile():
@@ -127,22 +113,7 @@ def test_PopSTRRightFile():
     assert main(args)==0
     args.verbose = True
     assert main(args)==0
-
     
-# test unzipped, unindexed VCFs return 1
-def test_UnzippedUnindexedFile():
-    fname1 = os.path.join(MRGVCFDIR, "test_file_gangstr_unzipped1.vcf")
-    fname2 = os.path.join(MRGVCFDIR, "test_file_gangstr_unzipped2.vcf")
-    with pytest.raises(ValueError) as info:
-        LoadReaders([fname1, fname2])
-    assert "is bgzipped and indexed" in str(info.value)
-
-    fname1 = os.path.join(MRGVCFDIR, "test_file_gangstr_unindexed1.vcf.gz")
-    fname2 = os.path.join(MRGVCFDIR, "test_file_gangstr_unindexed2.vcf.gz")
-    with pytest.raises(ValueError) as info:
-        LoadReaders([fname1, fname2])
-    assert "Could not find VCF index" in str(info.value)
-
 # test VCFs with different ref genome contigs return 1
 def test_WrongContigFile():
     args = base_argparse()
@@ -159,7 +130,6 @@ def test_WrongContigFile():
     with pytest.raises(ValueError) as info:
         main(args)
     assert "Different contigs found across VCF files." in str(info.value)
-
 
 def test_MissingFieldWarnings(capsys):
     args = base_argparse()
@@ -209,47 +179,6 @@ def test_GetInfoItem(capsys):
 
     retval = GetInfoItem(dummy_records, [True, True, False, False], 'END')
     assert retval == "END=120"
-
-def test_PrintCurrentRecords(capsys):
-    # Set up dummy class
-    class DummyRecordNoChrom:
-        def __init__(self, chrom, pos, ref, alts=[], info = {}):
-            self.POS = pos
-            self.REF = ref
-            self.ALTS = alts
-            self.INFO = info
-    # Set up dummy records
-    dummy_records = [] 
-    dummy_records.append(DummyRecord('chr1', 100, 'CAGCAG', info={'END': 120}))
-    dummy_records.append(DummyRecordNoChrom('chr1', 100, 'CAGCAG', info={'END': 120}))
-    
-    PrintCurrentRecords(dummy_records, [True, True])
-    captured = capsys.readouterr()
-    assert "Missing CHROM and POS in record" in captured.err
-
-def test_CheckMin():
-    assert CheckMin([True, False]) == False
-    with pytest.raises(ValueError) as info:
-        CheckMin([False, False])
-    assert "Unexpected error. Stuck in infinite loop and exiting." in str(info.value)
-
-def test_CheckVCFType():
-    snps_path = os.path.join(VCFDIR, "snps.vcf")
-    gangstr_path = os.path.join(VCFDIR, "test_gangstr.vcf")
-    hipstr_path = os.path.join(VCFDIR, "test_hipstr.vcf")
-    gangstr_vcf = vcf.Reader(filename=gangstr_path)
-    hipstr_vcf = vcf.Reader(filename=hipstr_path)
-    snps_vcf = vcf.Reader(filename=snps_path)
-    # TODO add tests to infer vcf type
-    assert "gangstr" == GetVCFType([gangstr_vcf], "gangstr")
-
-    with pytest.raises(ValueError) as info:
-        print(GetVCFType([gangstr_vcf, hipstr_vcf], "auto"))
-    assert "VCF files are of mixed types." in str(info.value)
-    
-    with pytest.raises(ValueError) as info:
-        print(GetVCFType([gangstr_vcf, snps_vcf], "auto"))
-    assert "Could not identify the type of this vcf" in str(info.value)
 
 def test_GetSampleInfo():
     # TODO add more in depth tests (Create a better dummy class or import from vcf files)
